Master

+++ view/forum/topics/detailsTopic.php
@@ -1,10 +1,18 @@
 <?php
+use App\Session;
     $topic = $result["data"]['topic']; 
     $posts = $result["data"]['posts']; 
 ?>
 
-<a href="index.php?ctrl=topic&action=deleteTopic&id=<?= $topic->getId() ?>">supprimer le topic</a>
-<a href="index.php?ctrl=topic&action=updateTopicForm&id=<?= $topic->getId() ?>">modifier le topic</a>
+<?php
+        if((Session::isAdmin()) || (Session::getUser()->getId() == $topic->getUser()->getId())){
+            ?>
+            <a href="index.php?ctrl=topic&action=deleteTopic&id=<?= $topic->getId() ?>">supprimer le topic</a>
+            <a href="index.php?ctrl=topic&action=updateTopicForm&id=<?= $topic->getId() ?>">modifier le topic</a>
+            <?php 
+        }
+?>
+
 
 <div id="form-container">
     <div id="form-header">
@@ -13,7 +21,7 @@
 
     <form id="form-content" action="index.php?ctrl=topic&action=createPost&id=<?= $topic->getId() ?>" method="post">
         <label for="content">contenu :</label>
-        <textarea id="content" type="text" name="content" cols="60" rows="5"></textarea>
+        <textarea id="content" type="text" name="content" cols="60" rows="5" required></textarea>
         <button type="submit" name ="submit">soumettre</button>
     </form>
 </div>
@@ -25,7 +33,16 @@
 if (!empty($posts)) {
 foreach($posts as $post ){ ?>
     <p>
-        <?= $post->getContent() ?> par <?= $post->getUser() ?> <a href="index.php?ctrl=topic&action=deletePost&id=<?= $post->getId() ?>">supprimer</a> <a href="index.php?ctrl=topic&action=updatePostForm&id=<?= $post->getId() ?>">modifier</a>
+        <?= $post->getContent() ?> par <?= $post->getUser() ?> 
+
+<?php
+        if((Session::isAdmin()) || (Session::getUser()->getId() == $post->getUser()->getId())){
+?>
+        <a href="index.php?ctrl=topic&action=deletePost&id=<?= $post->getId() ?>">supprimer</a> 
+        <a href="index.php?ctrl=topic&action=updatePostForm&id=<?= $post->getId() ?>">modifier</a>
+<?php 
+        }
+?>
     </p>
 <?php } 
 } else {
